backend/requirements.txt
agency-swarm==0.4.1
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.9.0
argcomplete==3.6.2
asyncpg==0.31.0
attrs==25.3.0
black==25.1.0
blinker==1.9.0
branca==0.8.2
cachetools==6.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
datamodel-code-generator==0.26.1
deepdiff==6.7.1
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
email_validator==2.2.0
fastapi==0.125.0
folium==0.20.0
genson==1.3.0
GeoAlchemy2==0.18.1
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflect==5.6.2
isort==5.13.2
Jinja2==3.1.6
jiter==0.10.0
jsonref==1.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mutagen==1.47.0
mypy_extensions==1.1.0
narwhals==1.42.1
numpy==2.3.0
openai==1.88.0
ordered-set==4.1.0
packaging==25.0
pandas==2.3.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
protobuf==6.31.1
psycopg2-binary==2.9.11
pyarrow==20.0.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rich==13.7.1
rpds-py==0.25.1
setuptools==78.1.1
shapely==2.1.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.45
srt==3.5.3
starlette==0.50.0
streamlit==1.46.0
streamlit-folium==0.25.3
tenacity==9.1.2
termcolor==2.4.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.38.0
vosk==0.3.44
websockets==15.0.1
wheel==0.45.1
xyzservices==2025.11.0
bcrypt==4.0.1
pgvector
passlib[bcrypt]
python-jose[cryptography]
